xserver libX11 libxcb mesa git update 15 Mar 2011

+++ xorg-server/include/propertyst.h
@@ -58,8 +58,8 @@ typedef struct _Property {
         struct _Property       *next;
 	ATOM 		propertyName;
 	ATOM		type;       /* ignored by server */
-	short		format;     /* format of data for swapping - 8,16,32 */
-	long		size;       /* size of data in (format/8) bytes */
+	uint32_t	format;     /* format of data for swapping - 8,16,32 */
+	uint32_t	size;       /* size of data in (format/8) bytes */
 	pointer         data;       /* private to client */
 	PrivateRec	*devPrivates;
 } PropertyRec;
